work in progress - got iceberg catalog initialization and create table to work
