+++ deltacat/catalog/__init__.py
@@ -2,8 +2,13 @@ from deltacat.catalog.model.properties import (  # noqa: F401
     CatalogProperties,
     get_catalog_properties,
 )
+from deltacat.catalog.model.catalog import ( # noqa: F401
+    Catalog, Catalogs
+)
 
 __all__ = [
     "CatalogProperties",
     "get_catalog_properties",
+    "Catalog",
+    "Catalogs"
 ]
