+++ deltacat/catalog/model/properties.py
@@ -10,14 +10,18 @@ from deltacat.utils.filesystem import resolve_path_and_filesystem
 def get_catalog_properties(
     *args,
     catalog: Optional[CatalogProperties] = None,
+    inner: Optional[CatalogProperties] = None,
     **kwargs,
 ) -> CatalogProperties:
     """
-    Helper function to fetch CatalogProperties instance.
+    Helper function to fetch CatalogProperties instance. You are meant to call this by providing your functions
+    kwargs, OR to directly pass through CatalogProperty configuration keys like "root" in kwargs.
 
-    This just passes through to CatalogProperties constructor if `catalog` not specified
+    This will look for a CatalogProperty value in the kwargs "catalog" or "inner". If these are found, it returns
+    the CatalogProperty value under that kwarg. Otherwise, it will pass through kwargs to the CatalogProperties
+    constructor.
     """
-    properties = catalog
+    properties = catalog if catalog is not None else inner
     if properties is not None and isinstance(properties, CatalogProperties):
         return properties
     elif properties is not None and not isinstance(properties, CatalogProperties):
