+++ dev-requirements.txt
@@ -9,6 +9,9 @@ memray == 1.6.0; platform_system != "Windows" and sys_platform != "darwin" and p
 moto == 4.1.12
 Pillow == 11.0.0
 pre-commit == 2.20.0
+pyiceberg[glue] >= 0.9.0
+pyiceberg[hive] >= 0.9.0
+pyiceberg[sql-sqlite] >= 0.9.0
 pyspark == 3.5.3
 pytest == 7.2.0
 pytest-benchmark == 4.0.0
