+++ source/index.rst
@@ -1,0 +1,20 @@
+.. deltacat documentation master file, created by
+   sphinx-quickstart on Thu Mar 27 10:59:30 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to deltacat's documentation!
+====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
